requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
google-generativeai==0.8.3
pdfplumber==0.11.4
gunicorn==22.0.0
Werkzeug==3.0.4
